+++ src/components/center_pannel/center_pannel.module.css
@@ -1,4 +1,3 @@
-
 .centerContainer {
     display: flex;
     flex-direction: column;
