+++ src/components/login_form/login_form.tsx
@@ -1,39 +1,89 @@
 import React from 'react'
 import styles from './login_form.module.css'
 
-interface Props {
-}
+interface Props {}
 
-function SidePannel({} : Props){
-    return (
+function LoginForm({} : Props){
     const loginContent = <> 
-                    <label className="radio-container">
-                        <input type="radio" name="option" className='login-radio' value="option1"/>
-                      <span className="radio-inner"></span>
-                      <span className="radio-label">Option 1</span>
+                <div className={styles.customCheckBoxHolder}>
+
+                    <input className={styles.customCheckBoxInput} name="option" id="cCB1" type="radio"/>
+                    <label className={styles.customCheckBoxWrapper} for="cCB1">
+                        <div className={styles.customCheckBox}>
+                            <div className="inner">Login</div>
+                        </div>
                     </label>
 
-                    <label className="radio-container">
-                      <input type="radio" className='login-radio' name="option" value="option2"/>
-                      <span className="radio-inner"></span>
-                      <span className="radio-label">Option 2</span>
+                    <input className={styles.customCheckBoxInput} name="option" id="cCB2" type="radio"/>
+                    <label className={styles.customCheckBoxWrapper} for="cCB2">
+                        <div className={styles.customCheckBox}>
+                            <div className="inner">Sign Up</div>
+                        </div>
                     </label>
+
+                </div>
+
                     <form>
-                        <input id='fname' className='input-box' placeholder="User name"/> <br/> 
-                        <input id='password' className='input-box' placeholder="Password"/> <br/> 
+                        <input id='lname' classNameName={styles.inputBox} placeholder="User name"/> <br/> 
+                        <input id='lpassword' classNameName={styles.inputBox} placeholder="Password"/> <br/> 
                     </form> 
 
                     <button onClick={LoginSubmit}> submit </button> 
-                    <h3>Sign up </h3>
-                    <form>
-                        <input id='fname' className='input-box' placeholder="User name"/> <br/> 
-                        <input id='password' className='input-box' placeholder="Password"/> <br/> 
-                    </form> 
-                    <button onClick={SignUpSubmit}> submit </button> 
                 </>;
 
-        return ( loginConent);
+        return (loginContent);
 }
 
-export default SidePannel;
 
+async function SignUpSubmit(){
+    const usernameInput = document.getElementById("fname") as HTMLInputElement;
+    const passwordInput = document.getElementById("password") as HTMLInputElement;
+
+    if (!usernameInput || !passwordInput){return(-1);}
+
+    const api_inputs = {
+            username: usernameInput.value,
+            password: passwordInput.value
+        };
+
+    const url = 'https://apichallenges.herokuapp.com/mirror/request';
+    const data = await fetch(
+        url, {
+            method: "POST", 
+            body: JSON.stringify(api_inputs)
+        }
+    );
+
+    if (data['status'] === 200){
+        console.log('You logged in');
+        window.location.href = "/home";
+    }
+    console.log(data);
+}
+
+async function LoginSubmit(){
+    const usernameInput = document.getElementById("fname") as HTMLInputElement;
+    const passwordInput = document.getElementById("password") as HTMLInputElement;
+
+    if (!usernameInput || !passwordInput){return(-1);}
+
+    const api_inputs = {
+            username: usernameInput.value,
+            password: passwordInput.value
+        };
+
+    const url = 'https://apichallenges.herokuapp.com/mirror/request';
+    const data = await fetch(
+        url, {
+            method: "POST", 
+            body: JSON.stringify(api_inputs)
+        }
+    );
+
+    if (data['status'] === 200){
+        console.log('You logged in');
+        window.location.href = "/home";
+    }
+    console.log(data);
+}
+export default LoginForm;
